feat: add wrong network notification
